ml_service/requirements.txt
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.7.1
psycopg2-binary==2.9.10
pgvector==0.3.6
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.0
huggingface-hub==0.24.5
